Bug fix in progress bar code
Previous version of progress bar code was dividing by 0 when the percent
complete was less than 1. Refactored the system to use the fraction
complete, rather than a percentage, and thus changed all the longints to
reals. This gets rid of the unnecessary divides by 100 as well (I wasn't
using the percentage value on screen anywhere so there's no reason to
keep it a percentage since 4D wants a value between 0 and 1).