Remove Plug-In Dependency (For Reals)
My previous change to remove the "MISC I" plug-in dependency from the
component didn't really work. You could compile the Host database but
the component itself did not compile. This change should actually fix
the problem, the plug-in command is called via EXECUTE FORMULA and only
if the plug-in is installed. Also the build method won't execute unless
the plug-in is installed.